working on breaking up the giant preview-directory file
