+++ frontend/app/view/preview/preview-directory-utils.tsx
@@ -1,0 +1,86 @@
+// Copyright 2025, Command Line Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { isBlank } from "@/util/util";
+import { Column } from "@tanstack/react-table";
+import dayjs from "dayjs";
+import React from "react";
+
+export const recursiveError = "recursive flag must be set for directory operations";
+export const overwriteError = "set overwrite flag to delete the existing file";
+export const mergeError = "set overwrite flag to delete the existing contents or set merge flag to merge the contents";
+
+export const displaySuffixes = {
+    B: "b",
+    kB: "k",
+    MB: "m",
+    GB: "g",
+    TB: "t",
+    KiB: "k",
+    MiB: "m",
+    GiB: "g",
+    TiB: "t",
+};
+
+export function getBestUnit(bytes: number, si: boolean = false, sigfig: number = 3): string {
+    if (bytes === undefined || bytes < 0) {
+        return "-";
+    }
+    const units = si ? ["kB", "MB", "GB", "TB"] : ["KiB", "MiB", "GiB", "TiB"];
+    const divisor = si ? 1000 : 1024;
+
+    let currentUnit = "B";
+    let currentValue = bytes;
+    let idx = 0;
+    while (currentValue > divisor && idx < units.length - 1) {
+        currentUnit = units[idx];
+        currentValue /= divisor;
+        idx += 1;
+    }
+
+    return `${parseFloat(currentValue.toPrecision(sigfig))}${displaySuffixes[currentUnit]}`;
+}
+
+export function getLastModifiedTime(unixMillis: number, column: Column<FileInfo, number>): string {
+    const fileDatetime = dayjs(new Date(unixMillis));
+    const nowDatetime = dayjs(new Date());
+
+    let datePortion: string;
+    if (nowDatetime.isSame(fileDatetime, "date")) {
+        datePortion = "Today";
+    } else if (nowDatetime.subtract(1, "day").isSame(fileDatetime, "date")) {
+        datePortion = "Yesterday";
+    } else {
+        datePortion = dayjs(fileDatetime).format("M/D/YY");
+    }
+
+    if (column.getSize() > 120) {
+        return `${datePortion}, ${dayjs(fileDatetime).format("h:mm A")}`;
+    }
+    return datePortion;
+}
+
+const iconRegex = /^[a-z0-9- ]+$/;
+
+export function isIconValid(icon: string): boolean {
+    if (isBlank(icon)) {
+        return false;
+    }
+    return icon.match(iconRegex) != null;
+}
+
+export function getSortIcon(sortType: string | boolean): React.ReactNode {
+    switch (sortType) {
+        case "asc":
+            return <i className="fa-solid fa-chevron-up dir-table-head-direction"></i>;
+        case "desc":
+            return <i className="fa-solid fa-chevron-down dir-table-head-direction"></i>;
+        default:
+            return null;
+    }
+}
+
+export function cleanMimetype(input: string): string {
+    const truncated = input.split(";")[0];
+    return truncated.trim();
+}
